chore(data-workflows): Extract listing-components into dedicated components
