+++ src/components/data-workflows/runs-collections-listings.ts
@@ -1,3 +1,17 @@
 import GbqToGcs from './gbq-to-gcs/runs/CollectionListing.vue';
+import GcsToGcs from './gcs-to-gcs/runs/CollectionListing.vue';
+import StorageToStorage from './storage-to-storage/runs/CollectionListing.vue';
+import StorageToTable from './storage-to-table/runs/CollectionListing.vue';
+import StorageToTables from './storage-to-tables/runs/CollectionListing.vue';
+import TableToStorage from './table-to-storage/runs/CollectionListing.vue';
+import TablesToTables from './tables-to-tables/runs/CollectionListing.vue';
+import VmLauncher from './vm-launcher/runs/CollectionListing.vue';
 
-export const GbqToGcsCollectionListing = GbqToGcs;
+export const GbqToGcsRunsCollectionListing = GbqToGcs;
+export const GcsToGcsRunsCollectionListing = GcsToGcs;
+export const StorageToStorageCollectionListing = StorageToStorage;
+export const StorageToTableRunsCollectionListing = StorageToTable;
+export const StorageToTablesRunsCollectionListing = StorageToTables;
+export const TableToStorageRunsCollectionListing = TableToStorage;
+export const TablesToTablesRunsCollectionListing = TablesToTables;
+export const VmLauncherRunsCollectionListing = VmLauncher;
