+++ src/components/data-workflows/storage-to-tables/runs/CollectionListing.vue
@@ -1,0 +1,96 @@
+<template>
+	<listing-component
+		:type="listingType"
+		:module-name="moduleName"
+		:headers="headers"
+		:overridden-columns="overriddenColumns"
+	>
+		<template v-slot:triggering_file="{ item: { id, triggering_file } }">
+			<router-link :to="{ name: routeName, params: { id } }">
+				<span class="font-weight-medium">{{ triggering_file }}</span>
+			</router-link>
+		</template>
+		<template v-slot:configuration_context.source="{ item: { configuration_context: { source } } }">
+			{{ source ? (source.type === 'gcs' ? source.gcs_source_bucket : source.sftp_host) : 'waiting...' }}
+		</template>
+
+		<template v-slot:dag_execution_date="{ item: { dag_execution_date } }">
+			{{ dag_execution_date | moment('YYYY/MM/DD - HH:mm') }}
+		</template>
+	</listing-component>
+</template>
+
+<script lang="ts">
+import { Component, Mixins } from 'vue-property-decorator';
+
+import RunCollectionMixin from '@/mixins/data-workflows/collection/run-collection-mixin';
+import ListingComponent from '@/components/data-workflows/common/listing/ListingComponent.vue';
+
+import { storageToTablesRuns } from '@/store/modules/easy-firestore/storage-to-tables-runs';
+import { STORAGE_TO_TABLES_RUNS_ITEM } from '@/constants/router/routes-names';
+
+@Component({
+	components: { ListingComponent }
+})
+export default class TestCollectionListing extends Mixins(RunCollectionMixin) {
+	moduleName: string = storageToTablesRuns.moduleName;
+	overriddenColumns: string[] = ['triggering_file', 'configuration_context.source', 'dag_execution_date'];
+
+	get routeName() {
+		return STORAGE_TO_TABLES_RUNS_ITEM;
+	}
+
+	get headers() {
+		return [
+			{
+				text: 'Account ID',
+				align: 'left',
+				sortable: true,
+				value: 'account'
+			},
+			{
+				text: 'Environment',
+				align: 'left',
+				sortable: true,
+				value: 'environment'
+			},
+			{
+				text: 'Triggering File',
+				align: 'left',
+				sortable: true,
+				value: 'triggering_file'
+			},
+			{
+				text: 'Source Storage',
+				align: 'left',
+				sortable: true,
+				value: 'configuration_context.source'
+			},
+			{
+				text: 'Destinations',
+				align: 'left',
+				sortable: true,
+				value: 'configuration_context.destinations.length'
+			},
+			{
+				text: 'Status',
+				align: 'left',
+				sortable: true,
+				value: 'status'
+			},
+			{
+				text: 'Execution Date',
+				align: 'left',
+				sortable: true,
+				value: 'dag_execution_date'
+			},
+			{
+				text: 'Actions',
+				align: 'center',
+				sortable: false,
+				value: 'actions'
+			}
+		];
+	}
+}
+</script>
