+++ src/views/data-workflows/table-to-storage/runs/ListingView.vue
@@ -1,25 +1,7 @@
 <template>
 	<div>
 		<data-management-header :workflowName="workflowName" :tabsItems="tabsItems" />
-		<listing-component
-			:type="listingType"
-			:module-name="moduleName"
-			:headers="headers"
-			:overridden-columns="overriddenColumns"
-			sort-by="dag_execution_date"
-			:sort-desc="true"
-			show-airflow-action
-		>
-			<template v-slot:firestore_conf_doc_id="{ item: { id, firestore_conf_doc_id } }">
-				<router-link :to="{ name: routeName, params: { id } }">
-					<span class="font-weight-medium">{{ firestore_conf_doc_id }}</span>
-				</router-link>
-			</template>
-
-			<template v-slot:dag_execution_date="{ item: { dag_execution_date } }">
-				{{ dag_execution_date | moment('YYYY/MM/DD - HH:mm') }}
-			</template>
-		</listing-component>
+		<table-to-storage-runs-collection-listing />
 	</div>
 </template>
 
@@ -27,41 +9,11 @@
 import { Component, Mixins } from 'vue-property-decorator';
 
 import HeaderInfosMixin from '../header-infos';
-import RunCollectionMixin from '@/mixins/data-workflows/collection/run-collection-mixin';
+import CollectionMixin from '@/mixins/data-workflows/collection/collection-mixin';
+import { TableToStorageRunsCollectionListing } from '@/components/data-workflows/runs-collections-listings';
 
-import { tableToStorageRuns } from '@/store/modules/easy-firestore/table-to-storage-runs';
-import { TABLE_TO_STORAGE_RUNS_ITEM } from '@/constants/router/routes-names';
-import {
-	ACCOUNT,
-	ACTIONS,
-	DAG_EXECUTION_DATE,
-	DESTINATION_BUCKET,
-	ENVIRONMENT,
-	FIRESTORE_CONF_DOC_ID,
-	OUTPUT_FILENAME,
-	STATUS
-} from '@/constants/data-workflows/listing/header-items';
-
-@Component
-export default class TableToStorageRunsListingView extends Mixins(HeaderInfosMixin, RunCollectionMixin) {
-	moduleName: string = tableToStorageRuns.moduleName;
-	overriddenColumns: string[] = ['firestore_conf_doc_id', 'dag_execution_date'];
-
-	get routeName() {
-		return TABLE_TO_STORAGE_RUNS_ITEM;
-	}
-
-	get headers() {
-		return [
-			ACCOUNT,
-			ENVIRONMENT,
-			FIRESTORE_CONF_DOC_ID,
-			DESTINATION_BUCKET,
-			OUTPUT_FILENAME,
-			STATUS,
-			DAG_EXECUTION_DATE,
-			ACTIONS
-		];
-	}
-}
+@Component({
+	components: { TableToStorageRunsCollectionListing }
+})
+export default class TableToStorageRunsListingView extends Mixins(HeaderInfosMixin, CollectionMixin) {}
 </script>
