+++ src/views/data-workflows/tables-to-tables/runs/ListingView.vue
@@ -1,25 +1,7 @@
 <template>
 	<div>
 		<data-management-header :workflowName="workflowName" :tabsItems="tabsItems" />
-		<listing-component
-			:type="listingType"
-			:module-name="moduleName"
-			:headers="headers"
-			:overridden-columns="overriddenColumns"
-			sort-by="dag_execution_date"
-			:sort-desc="true"
-			show-airflow-action
-		>
-			<template v-slot:dag_id="{ item: { id, dag_id } }">
-				<router-link :to="{ name: routeName, params: { id } }">
-					<span class="font-weight-medium">{{ dag_id }}</span>
-				</router-link>
-			</template>
-
-			<template v-slot:dag_execution_date="{ item: { dag_execution_date } }">
-				{{ dag_execution_date | moment('YYYY/MM/DD - HH:mm') }}
-			</template>
-		</listing-component>
+		<tables-to-tables-runs-collection-listing />
 	</div>
 </template>
 
@@ -27,31 +9,11 @@
 import { Component, Mixins } from 'vue-property-decorator';
 
 import HeaderInfosMixin from '../header-infos';
-import RunCollectionMixin from '@/mixins/data-workflows/collection/run-collection-mixin';
+import CollectionMixin from '@/mixins/data-workflows/collection/collection-mixin';
+import { TablesToTablesRunsCollectionListing } from '@/components/data-workflows/runs-collections-listings';
 
-import { getGbqToGbqRuns } from '@/store/modules/easy-firestore/get-gbq-to-gbq-runs';
-import { TABLES_TO_TABLES_RUNS_ITEM } from '@/constants/router/routes-names';
-import {
-	ACCOUNT,
-	ACTIONS,
-	DAG_EXECUTION_DATE,
-	DAG_ID,
-	ENVIRONMENT,
-	STATUS,
-	WORKFLOW_LENGTH
-} from '@/constants/data-workflows/listing/header-items';
-
-@Component
-export default class TablesToTablesRunsListingView extends Mixins(HeaderInfosMixin, RunCollectionMixin) {
-	moduleName: string = getGbqToGbqRuns.moduleName;
-	overriddenColumns: string[] = ['dag_id', 'dag_execution_date'];
-
-	get routeName() {
-		return TABLES_TO_TABLES_RUNS_ITEM;
-	}
-
-	get headers() {
-		return [ACCOUNT, ENVIRONMENT, DAG_ID, WORKFLOW_LENGTH, STATUS, DAG_EXECUTION_DATE, ACTIONS];
-	}
-}
+@Component({
+	components: { TablesToTablesRunsCollectionListing }
+})
+export default class TablesToTablesRunsListingView extends Mixins(HeaderInfosMixin, CollectionMixin) {}
 </script>
